requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
Brotli==1.1.0
cached-property==1.5.2
certifi==2025.6.15
cffi==1.15.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
distro==1.9.0
exceptiongroup==1.3.0
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.6
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.0.5
openai==1.39.0
Pillow==9.5.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydyf==0.9.0
Pygments==2.17.2
pyphen==0.14.0
python-dotenv==0.21.1
requests==2.31.0
rich==13.8.1
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
tabulate==0.9.0
tinycss2==1.2.1
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.7.1
urllib3==2.0.7
weasyprint==60.2
webencodings==0.5.1
yarl==1.9.4
zipp==3.15.0
zopfli==0.2.2
